Рефакторинг кода
