+++ src/main/java/school/faang/Main.java
@@ -1,55 +1,18 @@
 package school.faang;
 
-import lombok.Data;
+import school.faang.heroes.Archer;
+import school.faang.heroes.Warrior;
 
 public class Main {
-    @Data
-    public abstract class Character {
-        private String name;
-        private Integer power;
-        private Integer agility;
-        private Integer intellect;
-        private Integer health = 100;
+    public static void main(String[] args) {
+        Warrior warrior = new Warrior("Ivan");
+        Archer archer = new Archer("Sergey");
 
-        public Character(String name) {
-            this.name = name;
-        }
-
-        public Character(String name, Integer power, Integer agility, Integer intellect) {
-            this.name = name;
-            this.power = power;
-            this.agility = agility;
-            this.intellect = intellect;
-        }
-
-        public void takeDamage(Integer damage) {
-            this.health -= damage;
-        }
-
-        abstract void attack(Character character);
-    }
-
-    public class Warrior extends Character {
-
-        public Warrior(String name) {
-            super(name, 10, 5, 3);
-        }
-
-        @Override
-        public void attack(Character character) {
-            character.takeDamage(this.getPower());
-        }
-    }
-
-    public class Archer extends Character {
-
-        public Archer(String name) {
-            super(name, 3, 10, 5);
-        }
-
-        @Override
-        public void attack(Character character) {
-            character.takeDamage(this.getAgility());
-        }
+        System.out.printf("Здоровье %s %d\n", archer.getName(), archer.getHealth());
+        System.out.printf("Здоровье %s %d\n", warrior.getName(), warrior.getHealth());
+        warrior.attack(archer);
+        archer.attack(warrior);
+        System.out.printf("Здоровье %s %d\n", archer.getName(), archer.getHealth());
+        System.out.printf("Здоровье %s %d\n", warrior.getName(), warrior.getHealth());
     }
 }
